fixed a bug: when user go to his profile from other users page he doesnt see his own avatar
