+++ src/lib/getRandomPunishment.js
@@ -1,0 +1,7 @@
+let getRandomPunishment = (punishmentsArr) => {
+    let randomNum = Math.round( Math.random()* (punishmentsArr.length - 1));
+    return punishmentsArr[randomNum];
+}
+
+
+export default getRandomPunishment;
